+++ public/index.html
@@ -7,6 +7,7 @@
 <body>
   <img id="curWeapon"/>
   <div id="app"></div>
+  <p id="the-beg"></p>
   <script src="bundle.js"></script>
 </body>
 </html>
